保险添加回访记录   分配

+++ src/test/groovy/org/crygier/graphql/mlshop/model/Insurance.groovy
@@ -38,7 +38,7 @@ public class Insurance extends BosEntity{
     String customerTel;
 
     @SchemaDocumentation("返点")
-    Long rebate;
+    BigDecimal rebate;
 
     @SchemaDocumentation("返点时间")
     Long rebateTime;
